requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-jazzmin==3.0.0
idna==3.7
lxml==5.3.0
pycparser==2.22
PyNFe==0.4.4
pyOpenSSL==24.2.1
requests==2.32.3
signxml==3.2.2
sqlparse==0.5.1
urllib3==2.2.2
